Update ReplaceCommand.cs
Added an extra parameter to its constructor, a string, called _shape, which tells the command what kind of shape to replace it with, a cylinder or a cube. In the Execute() method, an if-else statement is added to distinguish between cylinder or cube (since before it could only do cylinder) based on the string _shape.