Update MainWindow.xaml
Before, under <MenuItem Header="Replace with...">, only the cylinder menu item had additional code that linked it to ReplaceWithCylinderOption from MainViewModel.cs, but the cube menu item had nothing else; that is the reason why clicking on Replace with... Cube in the app did nothing. I simply added the link with ReplaceWithCubeOption and that fixed it.