Update MainViewModel.cs
Modified the code of ReplaceWithCylinderOption and ReplaceWithCubeOption to modify the name of the mesh upon execution of these functions in order to distinguish them in the application's hierarchy. The cube option has also been modified to take into account the smallest dimensions of the original mesh when creating the rectangular prism that will replace it by invoking _sceneManager.GetSmallestMeshDimensions() instead.